requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.7
channels-api==0.4.1
channels-redis==2.3.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
daphne==2.2.5
dash==0.35.1
dash-core-components==0.42.1
dash-html-components==0.13.4
dash-renderer==0.17.0
dash-table==3.1.11
decorator==4.3.2
Django==2.1.5
django-filter==2.1.0
django-plotly-dash==0.9.8
django-redis==4.10.0
djangorestframework==3.9.1
dpd-components==0.1.0
Flask==1.0.2
Flask-Compress==1.4.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Markdown==3.0.1
MarkupSafe==1.1.0
msgpack==0.6.1
nbformat==4.4.0
plotly==3.5.0
PyHamcrest==1.9.0
pytz==2018.9
redis==3.1.0
requests==2.21.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
Werkzeug==0.14.1
zope.interface==4.6.0
